Fixed Cad Coordinates type

+++ src/Modules/Files/CustomCADs.Files.Domain/Cads/CadConstants.cs
@@ -4,7 +4,7 @@ public static class CadConstants
 {
     public static class Coordinates
     {
-        public const int CoordMin = -1000;
-        public const int CoordMax = 1000;
+        public const decimal CoordMin = -1000;
+        public const decimal CoordMax = 1000;
     }
 }
