+++ src/Shared/CustomCADs.Shared.Core/Common/ValueObjects/Coordinates.cs
@@ -1,3 +1,7 @@
 ﻿namespace CustomCADs.Shared.Core.Common.ValueObjects;
 
-public record Coordinates(int X = 0, int Y = 0, int Z = 0);
+public record Coordinates(
+    decimal X = 0,
+    decimal Y = 0,
+    decimal Z = 0
+);
