+++ test/UnitTests/Files/CustomCADs.UnitTests.Files.Domain/Cads/Create/Normal/CadCreateData.cs
@@ -1,3 +1,3 @@
 ﻿namespace CustomCADs.UnitTests.Files.Domain.Cads.Create.Normal;
 
-public class CadCreateData : TheoryData<string, string, int, int, int>;
+public class CadCreateData : TheoryData<string, string, decimal, decimal, decimal>;
